Moved kitti to utils
